Refactor QtPlot into PlotWidget directory and MVP
Re #34107
