+++ qt/widgets/plotting/CMakeLists.txt
@@ -6,15 +6,15 @@ set(MPL_SRC_FILES
     src/Mpl/ExternalPlotter.cpp
     src/Mpl/PeakPicker.cpp
     src/Mpl/PreviewPlot.cpp
-    src/Mpl/QtPlot.cpp
     src/Mpl/RangeSelector.cpp
     src/Mpl/SingleSelector.cpp
+    src/PlotWidget/QtPlotView.cpp
 )
 
 set(MPL_MOC_FILES
     inc/MantidQtWidgets/Plotting/Mpl/ContourPreviewPlot.h inc/MantidQtWidgets/Plotting/Mpl/PeakPicker.h
-    inc/MantidQtWidgets/Plotting/Mpl/PreviewPlot.h inc/MantidQtWidgets/Plotting/Mpl/QtPlot.h
     inc/MantidQtWidgets/Plotting/Mpl/RangeSelector.h inc/MantidQtWidgets/Plotting/Mpl/SingleSelector.h
+    inc/MantidQtWidgets/Plotting/Mpl/PreviewPlot.h inc/MantidQtWidgets/Plotting/PlotWidget/QtPlotView.h
 )
 
 # Include files aren't required, but this makes them appear in Visual Studio
@@ -25,11 +25,11 @@ set(MPL_INC_FILES
     inc/MantidQtWidgets/Plotting/ContourPreviewPlot.h
     inc/MantidQtWidgets/Plotting/PeakPicker.h
     inc/MantidQtWidgets/Plotting/PreviewPlot.h
-    inc/MantidQtWidgets/Plotting/QtPlot.h
+    inc/MantidQtWidgets/Plotting/PlotWidget/QtPlotView.h
 )
 
 set(MPL_UI_FILES inc/MantidQtWidgets/Plotting/ContourPreviewPlot.ui inc/MantidQtWidgets/Plotting/PreviewPlot.ui
-                 inc/MantidQtWidgets/Plotting/QtPlot.ui
+                 inc/MantidQtWidgets/Plotting/PlotWidget/PlotWidget.ui
 )
 
 mtd_add_qt_library(
@@ -48,7 +48,7 @@ mtd_add_qt_library(
   LINUX_INSTALL_RPATH "\$ORIGIN/../${LIB_DIR}"
 )
 
-set(TEST_FILES test/ContourPreviewPlotTest.h test/ExternalPlotterTest.h test/QtPlotTest.h)
+set(TEST_FILES test/ContourPreviewPlotTest.h test/ExternalPlotterTest.h test/PlotWidget/PlotPresenterTest.h)
 
 set(CXXTEST_EXTRA_HEADER_INCLUDE "${CMAKE_CURRENT_LIST_DIR}/test/PlottingTestInitialization.h")
 # Testing target
