+++ qt/widgets/plotting/inc/MantidQtWidgets/Plotting/PlotWidget/PlotWidget.ui
