+++ qt/widgets/plotting/inc/MantidQtWidgets/Plotting/PlotWidget/QtPlotView.h
