--- qt/widgets/plotting/inc/MantidQtWidgets/Plotting/QtPlot.h
@@ -1,11 +1,0 @@
-// Mantid Repository : https://github.com/mantidproject/mantid
-//
-// Copyright &copy; 2022 ISIS Rutherford Appleton Laboratory UKRI,
-//   NScD Oak Ridge National Laboratory, European Spallation Source,
-//   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
-// SPDX - License - Identifier: GPL - 3.0 +
-#pragma once
-
-#include <QtGlobal>
-
-#include "MantidQtWidgets/Plotting/Mpl/QtPlot.h"
